Added some files.
City list is a list of cities with their corresponding ID.  This is provided by the OpenWeatherMap API.
getWeather.py is a script I wrote that utilizes their API and my API key to return weather data.  I redirected the output from stdout to weather.txt.  Server.py is how I tested the Flask server, and it's close to what we'll be using for the actual thing, minus the visual stuff.